// src/constant/enums.ts
export enum GenreParams {
  ALL = '',
  ARPG = 'argp',
  MMOARPG = 'mmoarpg',
  STRATEGY = 'strategy',
  MMORPG = 'mmorpg',
  FIGHTING = 'fighting',
  ACTION_RPG = 'action_rpg',
  BATTLE_ROYALE = 'battle_royale',
  MOBA = 'moba',
  CARD = 'card',
  SPORTS = 'sports',
  RACING = 'racing',
  CARD_GAME = 'card_game',
  MMO = 'mmo',
  SOCIAL = 'social',
  FANTASY = 'fantasy',
}
export enum FilterParams {
  TITLE = 'title',
  DATE = 'release_date',
  PLATFORM = 'platform',
  DEV = 'developer',
}